course finder extracted to domain service
